requirements.txt
colorama==0.4.4
loguru==0.5.3
pyasn1==0.4.8
pycryptodome==3.12.0
PyQt5==5.15.6
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
PyQt5-stubs==5.15.2.0
